safari-600.7-branch/LayoutTests/editing/execCommand/break-non-editable-blockquote.html

<!DOCTYPE html>
<html>
<body>
<blockquote type="cite">
<div id="test" contenteditable><blockquote type="cite">hello world</blockquote></div>
</blockquote>
<script src="../../resources/dump-as-markup.js"></script>
<script>

var test = document.getElementById('test');
window.getSelection().setPosition(test.firstChild.firstChild, 5);
document.execCommand('InsertNewlineInQuotedContent', false, null);

Markup.description('This test ensures WebKit properly inserts newline in quoted contents, which is enclosed by a non-editable blockquote.');
Markup.dump(test);

</script>
</body>
</html>
