third_party/blink/web_tests/editing/execCommand/break-non-editable-blockquote-expected.txt

This test ensures WebKit properly inserts newline in quoted contents, which is enclosed by a non-editable blockquote.
| <blockquote>
|   type="cite"
|   "hello"
| <br>
| <blockquote>
|   type="cite"
|   " world"
